ai_model/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
numpy>=1.21.0
Pillow>=9.0.0